+++ src/legacy/core_plugins/metrics/index.ts
@@ -18,22 +18,25 @@
  */
 
 import { resolve } from 'path';
+import { Legacy } from 'kibana';
+import { PluginInitializerContext } from 'src/core/server';
+import { CoreSetup } from 'src/core/server';
+import { plugin } from './server/';
+import { CustomCoreSetup } from './server/plugin';
 
-import { fieldsRoutes } from './server/routes/fields';
-import { visDataRoutes } from './server/routes/vis';
-import { SearchStrategiesRegister } from './server/lib/search_strategies/search_strategies_register';
-
-export default function(kibana) {
+// eslint-disable-next-line import/no-default-export
+export default function(kibana: any) {
   return new kibana.Plugin({
-    require: ['kibana', 'elasticsearch'],
+    id: 'metrics',
+
+    require: ['kibana', 'elasticsearch', 'visualizations', 'data'],
 
     uiExports: {
-      visTypes: ['plugins/metrics/kbn_vis_types'],
-      interpreter: ['plugins/metrics/tsvb_fn'],
       styleSheetPaths: resolve(__dirname, 'public/index.scss'),
+      hacks: ['plugins/metrics/'],
     },
 
-    config(Joi) {
+    config(Joi: any) {
       return Joi.object({
         enabled: Joi.boolean().default(true),
         chartResolution: Joi.number().default(150),
@@ -41,11 +44,11 @@ export default function(kibana) {
       }).default();
     },
 
-    init(server) {
-      fieldsRoutes(server);
-      visDataRoutes(server);
+    init(server: Legacy.Server) {
+      const initializerContext = {} as PluginInitializerContext;
+      const core = { http: { server } } as CoreSetup & CustomCoreSetup;
 
-      SearchStrategiesRegister.init(server);
+      plugin(initializerContext).setup(core);
     },
   });
 }
