+++ src/legacy/core_plugins/metrics/public/editor_controller.js
@@ -20,11 +20,10 @@
 import React from 'react';
 import { render, unmountComponentAtNode } from 'react-dom';
 import { I18nContext } from 'ui/i18n';
-import chrome from 'ui/chrome';
-import { fetchIndexPatternFields } from '../lib/fetch_fields';
+import { fetchIndexPatternFields } from './lib/fetch_fields';
 
-function ReactEditorControllerProvider(config) {
-  class ReactEditorController {
+export function createEditorController(config, savedObjectsClient) {
+  return class EditorController {
     constructor(el, savedObj) {
       this.el = el;
 
@@ -36,7 +35,6 @@ function ReactEditorControllerProvider(config) {
     }
 
     fetchDefaultIndexPattern = async () => {
-      const savedObjectsClient = chrome.getSavedObjectsClient();
       const indexPattern = await savedObjectsClient.get(
         'index-pattern',
         config.get('defaultIndex')
@@ -85,12 +83,5 @@ function ReactEditorControllerProvider(config) {
     destroy() {
       unmountComponentAtNode(this.el);
     }
-  }
-
-  return {
-    name: 'react_editor',
-    handler: ReactEditorController,
   };
 }
-
-export { ReactEditorControllerProvider };
